Funcionalidades Store Carrinho concluidas

+++ biblioteca/src/store/carrinho.js
@@ -18,59 +18,57 @@ export const carrinhoStore = defineStore('carrinho', {
             return Promise.resolve()
         },
         async addItem(produto) {
+            const produtoExiste = this.itens.find(item => item.id === produto.id)
             
-            if(this.itens.find(item => item.id === produto.id)) {
-                this.itens.find(item => item.id === produto.id).quantidade++
-                const{data} = await axios.post(`http://localhost:3000/carrinho/${produto.id}`,produto)
+            if(produtoExiste && produtoExiste.quantidade < produto.estoque) {
+                    produtoExiste.quantidade++
+                    const{data} = await axios.put(`http://localhost:3000/carrinho/${produto.id}`, produtoExiste)
                 return Promise.resolve()
             }
+            produto.quantidade = 1
             const{data} = await axios.post(`http://localhost:3000/carrinho/`,produto)
-            data.quantidade = 1
             this.itens.push(data)
             this.total += data.preco
             return Promise.resolve()
             
-            // for(let i = 0; i < this.itens.length; i++) {
-            //     if(this.itens[i].quantidade === produto.estoque) {
-            //         return
-            //     }
-            //     if(this.itens[i].livro === produto.livro ) {
-            //         this.itens[i].quantidade++
-            //         this.itens[i].total = this.itens[i].quantidade * this.itens[i].preco
-            //         this.total += this.itens[i].preco
-            //         return
-            //     }
-            // }
-            // this.itens.push(produto)
-            // this.total += produto.preco
         },
-        async comprar(livros) {
-            await axios.put(`http://localhost:3000/livros/${livros.id}`,livros)
-            const index = this.itens.findIndex(item => item.id === livros.id)
+        async comprar() {
+            this.getItens()
 
-            this.itens.splice(index, 1, { ...livros })
+            this.itens.map(async item => {
+                const {data} = await axios.get(`http://localhost:3000/livros/${item.id}`)
+                if(item.quantidade > data.quantidade) {
+                    alert('Não temos essa quantidade em estoque')
+                    return Promise.resolve()
+                }
+                data.quantidade -= item.quantidade
+                await axios.put(`http://localhost:3000/livros/${item.id}`, data)
+                //this.removeTudo()
+                }
+            )
+            return Promise.resolve()
+        
         },
         async removeItem(produto) {
-            const{data} = await axios.delete(`http://localhost:3000/carrinho/${produto.id}`)
-            this.itens.splice(produto.id, 1)
+            const produtoExiste = this.itens.find(item => item.id === produto.id)
+
+            if(produtoExiste.quantidade > 1) {
+                produtoExiste.quantidade--
+                const{dados} = await axios.put(`http://localhost:3000/carrinho/${produto.id}`, produtoExiste)
+            } else {
+                const{data} = await axios.delete(`http://localhost:3000/carrinho/${produto.id}`)
+                this.itens.splice(produto.id, 1)
+            }
             this.total -= produto.preco
             return Promise.resolve()
 
-            // for(let i = 0; i < this.itens.length; i++) {
-            //     if(this.itens[i].livro === produto) {
-            //         this.itens[i].quantidade--
-            //         this.itens[i].total = this.itens[i].quantidade * this.itens[i].preco
-            //         this.total -= this.itens[i].preco
-            //         if(this.itens[i].quantidade === 0) {
-            //             this.itens.splice(i, 1)
-            //         }
-            //         return
-            //     }
-            // }
         },
-        removeTudo() {
-            this.itens = []
-            this.total = 0
+        async removeTudo() {
+           this.itens.map(async item => {
+                const{data} = await axios.delete(`http://localhost:3000/carrinho/${item.id}`)
+                this.itens.splice(item.id, 1)
+            }
+            )
         },
         
     },
